+++ packages/frontend/src/main/waggleManager.ts
@@ -1,54 +1,57 @@
 import waggle, { DataServer, ConnectionsManager } from '@quiet/waggle'
-import getPort from 'get-port'
+import logger from './logger'
+import { Command } from 'commander'
+const program = new Command()
 
-export const getPorts = async (): Promise<{
-  socksPort: number
-  libp2pHiddenService: number
-  controlPort: number
-  httpTunnelPort: number
-  dataServer: number
-}> => {
-  const controlPort = await getPort()
-  const httpTunnelPort = await getPort()
-  const socksPort = await getPort()
-  const libp2pHiddenService = await getPort()
-  const dataServer = await getPort()
-  return {
-    socksPort,
-    libp2pHiddenService,
-    controlPort,
-    httpTunnelPort,
-    dataServer
-  }
-}
+const log = logger('waggleManager')
 
-export type ApplicationPorts = Awaited<ReturnType<typeof getPorts>>
-export const runWaggle = async (
-  ports: ApplicationPorts,
-  appDataPath: string
-): Promise<{
+program
+  .option('-s, --socksPort <number>', 'Socks proxy port')
+  .option('-h, --httpTunnelPort <number>', 'Tor http tunnel port')
+  .option('-a, --appDataPath <string>', 'Path of application data directory')
+  .option('-c, --controlPort <number>', 'Tor control port')
+  .option('-d, --dataServerPort <number>', 'Socket io data server port')
+  .option('-r, --resourcesPath <string>', 'Application resources path')
+  .option('-l, --libp2pHiddenService <number>', 'Libp2p tor hidden service port')
+
+program.parse(process.argv)
+const options = program.opts()
+
+export const runWaggle = async (): Promise<{
   connectionsManager: ConnectionsManager
   dataServer: DataServer
 }> => {
-  const dataServer = new waggle.DataServer(ports.dataServer)
+  const dataServer = new waggle.DataServer(options.dataServerPort)
   await dataServer.listen()
 
   const isDev = process.env.NODE_ENV === 'development'
-  const resourcesPath = isDev ? null : process.resourcesPath
+  const resourcesPath = isDev ? null : options.resourcesPath.trim()
 
   const connectionsManager = new waggle.ConnectionsManager({
-    port: ports.libp2pHiddenService,
+    port: options.libp2pHiddenService,
     agentHost: 'localhost',
-    agentPort: ports.socksPort,
-    httpTunnelPort: ports.httpTunnelPort,
+    agentPort: options.socksPort,
+    httpTunnelPort: options.httpTunnelPort,
     io: dataServer.io,
     options: {
       env: {
-        appDataPath: `${appDataPath}/Quiet`,
+        appDataPath: `${options.appDataPath.trim()}/Quiet`,
         resourcesPath
       },
       spawnTor: true,
-      torControlPort: ports.controlPort
+      torControlPort: options.controlPort
+    }
+  })
+
+  process.on('message', async (message) => {
+    if (message === 'close') {
+      try {
+        await dataServer.close()
+        await connectionsManager.closeAllServices()
+      } catch (e) {
+        log.error('Error occured while closing waggle services', e)
+      }
+      process.send('closed-services')
     }
   })
 
@@ -59,4 +62,7 @@ export const runWaggle = async (
 
 export const waggleVersion = waggle.version
 
-export default { getPorts, runWaggle, waggleVersion }
+runWaggle().catch(e => {
+  log.error('Error occurred while initializing waggle', e)
+  throw Error(e.message)
+})