+++ packages/frontend/src/renderer/components/widgets/channels/NestedMessageContent.tsx
@@ -1,6 +1,8 @@
-import { Grid, makeStyles, Typography } from '@material-ui/core'
-import { DisplayableMessage } from '@quiet/nectar'
 import React from 'react'
+import theme from '../../../theme'
+import { Grid, makeStyles, Typography } from '@material-ui/core'
+import { DisplayableMessage, SendingStatus } from '@quiet/nectar'
+import classNames from 'classnames'
 
 const useStyles = makeStyles(() => ({
   message: {
@@ -15,22 +17,33 @@ const useStyles = makeStyles(() => ({
   },
   nextMessage: {
     paddingTop: 4
+  },
+  pending: {
+    color: theme.palette.colors.lightGray
   }
 }))
 
 export interface NestedMessageContentProps {
   message: DisplayableMessage
+  pending: boolean
   index: number
 }
 
-export const NestedMessageContent: React.FC<NestedMessageContentProps> = ({ message, index }) => {
+export const NestedMessageContent: React.FC<NestedMessageContentProps> = ({ message, pending, index }) => {
   const classes = useStyles({})
 
   const outerDivStyle = index > 0 ? classes.nextMessage : classes.firstMessage
 
   return (
     <Grid item className={outerDivStyle}>
-      <Typography className={classes.message} data-testid={`messagesGroupContent-${message.id}`}>{message.message}</Typography>
+      <Typography
+        className={classNames({
+          [classes.message]: true,
+          [classes.pending]: pending
+        })}
+        data-testid={`messagesGroupContent-${message.id}`}>
+        {message.message}
+      </Typography>
     </Grid>
   )
 }