+++ packages/nectar/src/sagas/messages/messages.master.saga.ts
@@ -4,11 +4,13 @@ import { all } from 'typed-redux-saga'
 import { publicChannelsActions } from '../publicChannels/publicChannels.slice'
 import { messagesActions } from './messages.slice'
 import { sendMessageSaga } from './sendMessage/sendMessage.saga'
+import { updateMessagesSendingStatusSaga } from './updateMessagesSendingStatus/updateMessagesSendingStatus.saga'
 import { verifyMessagesSaga } from './verifyMessage/verifyMessages.saga'
 
 export function* messagesMasterSaga(socket: Socket): Generator {
   yield all([
     takeEvery(messagesActions.sendMessage.type, sendMessageSaga, socket),
-    takeEvery(publicChannelsActions.incomingMessages.type, verifyMessagesSaga)
+    takeEvery(publicChannelsActions.incomingMessages.type, verifyMessagesSaga),
+    takeEvery(publicChannelsActions.incomingMessages.type, updateMessagesSendingStatusSaga)
   ])
 }