+++ packages/nectar/src/sagas/messages/messages.selectors.ts
@@ -1,7 +1,8 @@
 import { createSelector } from 'reselect'
 import { StoreKeys } from '../store.keys'
 import { CreatedSelectors, StoreState } from '../store.types'
-import { messageVerificationStatusAdapter } from './verifyMessage/verifyMessageAdapter'
+import { messageSendingStatusAdapter, messageVerificationStatusAdapter } from './messages.adapter.ts'
+import { MessageSendingStatus, SendingStatus } from './messages.types'
 
 const messagesSlice: CreatedSelectors[StoreKeys.Messages] = (state: StoreState) =>
   state[StoreKeys.Messages]
@@ -17,7 +18,14 @@ export const messagesVerificationStatus = createSelector(messagesSlice, reducerS
     .selectEntities(reducerState.messageVerificationStatus)
 )
 
+export const messagesSendingStatus = createSelector(messagesSlice, reducerState => {
+  return messageSendingStatusAdapter
+    .getSelectors()
+    .selectEntities(reducerState.messageSendingStatus)
+})
+
 export const messagesSelectors = {
   publicKeysMapping,
-  messageVerificationStatusAdapter
+  messagesVerificationStatus,
+  messagesSendingStatus
 }