+++ packages/nectar/src/sagas/publicChannels/publicChannels.slice.ts
@@ -18,6 +18,7 @@ import {
   ChannelMessagesIdsResponse,
   SubscribeToTopicPayload,
   AskForMessagesPayload,
+  PendingMessage,
   IncomingMessages,
   MarkUnreadMessagesPayload,
   ClearUnreadMessagesPayload,