+++ packages/nectar/src/utils/tests/factories.ts
@@ -11,7 +11,7 @@ import {
 import { getCrypto } from 'pkijs'
 import { stringToArrayBuffer } from 'pvutils'
 import { keyObjectFromString, verifySignature } from '@quiet/identity'
-import { MessageType } from '../../sagas/messages/messages.types'
+import { MessageType, SendingStatus } from '../../sagas/messages/messages.types'
 import { DateTime } from 'luxon'
 
 export const getFactory = async (store: Store) => {
@@ -198,6 +198,11 @@ export const getFactory = async (store: Store) => {
     code: 500,
     message: 'Community error',
     community: factory.assoc('Community', 'id')
+  })
+
+  factory.define('MessageSendingStatus', messages.actions.addMessagesSendingStatus, {
+    id: factory.assoc('Message', 'id'),
+    status: SendingStatus.Pending
   })
 
   return factory