+++ packages/identity/CHANGELOG.md
@@ -3,6 +3,22 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [0.3.2-alpha.0](https://github.com/ZbayApp/identity/compare/@quiet/identity@0.3.0-alpha.11...@quiet/identity@0.3.2-alpha.0) (2022-04-15)
+
+**Note:** Version bump only for package @quiet/identity
+
+
+
+
+
+## [0.3.1-alpha.0](https://github.com/ZbayApp/identity/compare/@quiet/identity@0.3.0-alpha.11...@quiet/identity@0.3.1-alpha.0) (2022-04-15)
+
+**Note:** Version bump only for package @quiet/identity
+
+
+
+
+
 # [0.3.0-alpha.11](https://github.com/ZbayApp/identity/compare/@quiet/identity@0.3.0-alpha.10...@quiet/identity@0.3.0-alpha.11) (2022-04-04)
 
 **Note:** Version bump only for package @quiet/identity